renamed file tracking if custom text is long
