--- frontend/src/ts/states/custom-text.ts
@@ -1,9 +1,0 @@
-let customTestName = ""; // It should be empty when the text is not saved or a saved text has been modified
-
-export function getCustomTextName(): string {
-  return customTestName;
-}
-
-export function setCustomTextName(newName: string): void {
-  customTestName = newName;
-}
